k8s changes
Former-commit-id: 118d94194bbe2a06493383663b4f9b4b6d5ca4e7
Former-commit-id: 4ee2c802b0c0685f8da7250da73be0d08577c91e [formerly fcc17260e4f422dd79de9db578863ee84de4c554]
Former-commit-id: 68a5c04c1750de696db2d2bb299b7d20bb57ec36

+++ src/Server/k8s/deploy.sh
@@ -1,23 +1,42 @@
 #!/bin/bash
 
+# trap ctrl-c and call ctrl_c()
+trap close INT
+
+if [ "$EUID" -ne 0 ]; then 
+  tput setaf 1 
+  tput bold 
+  echo "-------------------------------------------------------------------------------------------"
+  echo "Note this script needs sudo permissions... Please run it again with sudo."
+  echo "-------------------------------------------------------------------------------------------"
+  tput sgr 0 
+  exit
+fi
+
 echo "Deleting k8s resources..."
 
 kubectl delete deployments --all
 kubectl delete services --all
-kubectl delete configmap local-config|| true
+kubectl delete configmap local-config || true
+kubectl delete configmap internal-urls || true
+kubectl delete configmap endpoints || true
 kubectl delete -f ingress.yml
 
 echo "Deploying infrastructure components..."
-kubectl create -f pgsql-data.yml -f redis-data.yml -f config-map.yml
+kubectl create -f pgsql-data.yml -f redis-data.yml -f config-map.yml -f internal-urls.yml
 
 echo "Creating services..."
 kubectl create -f services.yml
 kubectl create -f ingress.yml
 
+tput bold
+tput setaf 2 # yellow
 ip_regex='([0-9]{1,3}\.){3}[0-9]{1,3}|[A-B]|[a-b]'
+echo "-------------------------------------------------------------------------------------------"
+echo "Please wait, retrieving ingress DNS/IP"
 while true; do
     printf "."
-    ingressUrl=$(kubectl get svc ingress-nginx -n ingress-nginx -o=jsonpath="{.status.loadBalancer.ingress[0].ip}" | kubectl get svc ingress-nginx -n ingress-nginx -o=jsonpath="{.status.loadBalancer.ingress[0].hostname}")
+    ingressUrl=$(kubectl get ing restaurant-ingress -o=jsonpath="{.status.loadBalancer.ingress[0].ip}")
     if [[ $ingressUrl =~ $ip_regex ]]; then
         break
     fi
@@ -28,12 +47,27 @@ printf "\n"
 externalURL=$ingressUrl
 echo "Using $externalURL as the external DNS/IP of the K8s cluster"
 
+hostfile="/etc/hosts"
+tput bold
+tput setaf 3 # yellow
+
+HOSTS=$(kubectl get ing restaurant-ingress | awk '{print $2}')
+for host in $(echo $HOSTS | tr "," "\n")
+do
+ if ! grep -q "^$externalURL $host" "$hostfile"; then
+    echo "- Adding entry '$externalURL $host"
+    echo "$externalURL $host" >> "$hostfile"
+ fi
+done
+
+tput bold
+tput setaf 0 # reset
+kubectl create configmap endpoints \
+    "--from-literal=basket_pub=http://api.restaurant.com/basket" \
+    "--from-literal=order_pub=http://api.restaurant.com/basket" \
+    "--from-literal=menu_pub=http://api.restaurant.com/basket" \
+    "--from-literal=identity_pub=http://api.restaurant.com/basket" \
+    "--from-literal=dashboard_pub=http://dashboard.restaurant.com" \
 
 kubectl create -f deployment.yml
-
-echo "Basket API is exposed at http://$externalURL/basket/swagger/index.html
-Order API at http://$externalURL/order/swagger-ui.html 
-Menu at http://$externalURL/menu/swagger/index.html
-Identity http://$externalURL/identity/swagger/index.html"
-
 echo "Deployment DONE!"
